Create Base Class to inherit
